Updated Category
Moved handler event to Expense
